requirements.txt
Scrapy
selenium
webdriver_manager
python-dotenv
selenium